+++ Web/RDD.Web.Tests/Models/MyValueObject.cs
@@ -1,12 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace RDD.Web.Tests.Models
+﻿namespace RDD.Web.Tests.Models
 {
     public class MyValueObject
     {
         public int Id { get; set; }
         public string Name { get; set; }
+        public User User { get; set; }
     }
 }