+++ Web/RDD.Web/Models/Metadata.cs
@@ -1,54 +1,17 @@
 ﻿using RDD.Domain;
-using RDD.Domain.Models.Querying;
 using System;
-using System.Collections.Generic;
 
 namespace RDD.Web.Models
 {
     public class Metadata
     {
         public MetadataHeader Header { get; set; }
-
         public object Data { get; set; }
 
-        public Metadata(object datas, Options options, Page page, IPrincipal principal)
+        public Metadata(object datas, IPrincipal principal, DateTime generatedAt)
         {
-            Header = new MetadataHeader(principal) { Generated = DateTime.Now };
+            Header = new MetadataHeader(principal) { Generated = generatedAt };
             Data = datas;
-
-            var offset = page.Offset;
-            var limit = page.Limit;
-            var count = page.TotalCount;
-
-            string next = null;
-            string previous = null;
-
-            if (offset + limit < count)
-            {
-                //var nextOffset = offset + limit;
-                //var url = HttpContext.Current.Request.Url;
-
-                //next = BuildUrlFromOffsetLimitAndQueryParams(url, nextOffset, limit, queryParameters);
-            }
-
-            if (offset > 0)
-            {
-                //var previousOffset = Math.Max(0, offset - limit);
-                //var url = HttpContext.Current.Request.Url;
-
-                //previous = BuildUrlFromOffsetLimitAndQueryParams(url, previousOffset, limit, queryParameters);
-            }
-
-            Header.Paging = new MetadataPaging() { count = count, offset = offset, limit = limit, next = next, previous = previous };
-        }
-
-        public Dictionary<string, object> ToDictionary()
-        {
-            return new Dictionary<string, object>
-            {
-                {"header", Header},
-                {"data", Data}
-            };
         }
     }
 }