+++ Web/RDD.Web/Serialization/Serializers/BaseClassSerializer.cs
@@ -1,17 +1,17 @@
-﻿using RDD.Domain.Helpers.Expressions;
+﻿using Newtonsoft.Json.Serialization;
+using RDD.Domain.Helpers.Expressions;
 using RDD.Web.Querying;
 using RDD.Web.Serialization.Providers;
 using RDD.Web.Serialization.Reflection;
-using RDD.Web.Serialization.Serializers;
 using RDD.Web.Serialization.UrlProviders;
 using System;
 
-namespace RDD.Web.Serialization
+namespace RDD.Web.Serialization.Serializers
 {
     public class BaseClassSerializer : EntitySerializer
     {
-        public BaseClassSerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, IUrlProvider urlProvider, Type workingType)
-            : base(serializerProvider, reflectionProvider, urlProvider, workingType) { }
+        public BaseClassSerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, NamingStrategy namingStrategy, IUrlProvider urlProvider, Type workingType)
+            : base(serializerProvider, reflectionProvider, namingStrategy, urlProvider, workingType) { }
 
         protected override IExpressionTree CorrectFields(object entity, IExpressionTree fields)
         {