+++ Web/RDD.Web/Serialization/Serializers/DateTimeSerializer.cs
@@ -1,5 +1,5 @@
-﻿using RDD.Domain.Helpers.Expressions;
-using RDD.Domain.Json;
+﻿using Newtonsoft.Json;
+using RDD.Domain.Helpers.Expressions;
 using RDD.Web.Serialization.Providers;
 using System;
 
@@ -9,18 +9,18 @@ namespace RDD.Web.Serialization.Serializers
     {
         public DateTimeSerializer(ISerializerProvider serializerProvider) : base(serializerProvider) { }
 
-        public override IJsonElement ToJson(object entity, IExpressionTree fields)
+        public override void WriteJson(JsonTextWriter writer, object entity, IExpressionTree fields)
         {
             switch (entity)
             {
-                case DateTime d: return ToJson(d, fields);
-                default: return base.ToJson(entity, fields);
-            }
-        }
+                case DateTime d:
+                    writer.WriteValue(DateTime.SpecifyKind(d, DateTimeKind.Unspecified));
+                    break;
 
-        public IJsonElement ToJson(DateTime entity, IExpressionTree fields)
-        {
-            return new JsonValue { Content = DateTime.SpecifyKind(entity, DateTimeKind.Unspecified) };
+                default:
+                    base.WriteJson(writer, entity, fields);
+                    break;
+            }
         }
     }
 }