+++ Web/RDD.Web/Serialization/Serializers/DictionarySerializer.cs
@@ -1,6 +1,7 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
 using RDD.Domain.Exceptions;
 using RDD.Domain.Helpers.Expressions;
-using RDD.Domain.Json;
 using RDD.Web.Serialization.Providers;
 using RDD.Web.Serialization.Reflection;
 using System.Collections;
@@ -10,47 +11,40 @@ namespace RDD.Web.Serialization.Serializers
 {
     public class DictionarySerializer : ObjectSerializer
     {
-        public DictionarySerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider) : base(serializerProvider, reflectionProvider, typeof(IDictionary)) { }
+        public DictionarySerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, NamingStrategy namingStrategy) 
+            : base(serializerProvider, reflectionProvider, namingStrategy, typeof(IDictionary)) { }
 
-        public override IJsonElement ToJson(object entity, IExpressionTree fields)
-        {
-            return ToJson(entity as IDictionary, fields);
-        }
+        public override void WriteJson(JsonTextWriter writer, object entity, IExpressionTree fields)
+            => WriteJson(writer, entity as IDictionary, fields);
 
-        protected IJsonElement ToJson(IDictionary dico, IExpressionTree fields)
+        protected void WriteJson(JsonTextWriter writer, IDictionary dico, IExpressionTree fields)
         {
+            writer.WriteStartObject();
+
             if (fields.Children.Any())
             {
-                return ToJsonWithFieldsFilter(dico, fields);
-            }
-
-            var result = new JsonObject();
-            foreach (var key in dico.Keys)
-            {
-                SerializeKvp(result, key.ToString(), dico[key], fields, null);
-            }
-
-            return result;
-        }
-
-        private IJsonElement ToJsonWithFieldsFilter(IDictionary dico, IExpressionTree fields)
-        {
-            var result = new JsonObject();
-            foreach (var child in fields.Children)
-            {
-                var concreteChild = child.Node as ItemExpression;
-                try
+                foreach (var child in fields.Children)
                 {
-                    var value = dico[concreteChild.Name];
-                    SerializeKvp(result, concreteChild.Name, value, child, null);
+                    var concreteChild = child.Node as ItemExpression;
+                    try
+                    {
+                        WriteKvp(writer, NamingStrategy.GetDictionaryKey(concreteChild.Name), dico[concreteChild.Name], child, null);
+                    }
+                    catch
+                    {
+                        throw new BadRequestException($"Unknown key {concreteChild.Name }");
+                    }
                 }
-                catch
+            }
+            else
+            {
+                foreach (var key in dico.Keys)
                 {
-                    throw new BadRequestException($"Unknown key {concreteChild.Name }");
+                    WriteKvp(writer, NamingStrategy.GetDictionaryKey(key.ToString()), dico[key], fields, null);
                 }
             }
 
-            return result;
+            writer.WriteEndObject();
         }
     }
 }