+++ Web/RDD.Web/Serialization/Serializers/ToStringSerializer.cs
@@ -1,5 +1,5 @@
-﻿using RDD.Domain.Helpers.Expressions;
-using RDD.Domain.Json;
+﻿using Newtonsoft.Json;
+using RDD.Domain.Helpers.Expressions;
 using RDD.Web.Serialization.Providers;
 
 namespace RDD.Web.Serialization.Serializers
@@ -8,7 +8,7 @@ namespace RDD.Web.Serialization.Serializers
     {
         public ToStringSerializer(ISerializerProvider serializerProvider) : base(serializerProvider) { }
 
-        public override IJsonElement ToJson(object entity, IExpressionTree fields)
-            => base.ToJson(entity.ToString(), fields);
+        public override void WriteJson(JsonTextWriter writer, object entity, IExpressionTree fields)
+            => writer.WriteValue(entity.ToString());
     }
 }